requirements.txt
langchain-openai>=0.2.0
python-dotenv>=1.0.0
pyshark>=0.6
tldextract>=5.1.0
